+++ modules/darwin/home-manager.nix
@@ -13,6 +13,7 @@
   qmd,
   nix-secrets,
   stylix,
+  direnv-instant,
   nixie-homes,
   ...
 }:
@@ -30,6 +31,7 @@ in
       nvf.homeManagerModules.default
       qmd.homeModules.default
       stylix.homeModules.stylix
+      direnv-instant.homeModules.direnv-instant
     ];
     extraSpecialArgs = { inherit nix-secrets; };
     users.${primaryUser} = {