+++ modules/common/github-secrets.nix
@@ -13,14 +13,14 @@ let
 in
 {
   age.secrets.github-ratelimit = {
-    file = "${nix-secrets}/github-ratelimit.age";
+    file = "${nix-secrets}/github/ratelimit.age";
     path = "${home}/.ssh/github-ratelimit";
     owner = primaryUser;
     mode = "0600";
   };
 
   age.secrets.github-ssh-key = {
-    file = "${nix-secrets}/github-ssh-key.age";
+    file = "${nix-secrets}/github/ssh-key.age";
     path = "${home}/.ssh/github-ssh-key";
     owner = primaryUser;
     mode = "0600";